Switch to a src layout
