+++ python-scripts/metadatavalidator/src/metadatavalidator/__init__.py
@@ -1,0 +1,2 @@
+
+from ._meta import __author__, __version__
