+++ python-scripts/metadatavalidator/src/metadatavalidator/_meta.py
@@ -1,0 +1,4 @@
+"""Version and author"""
+
+__version__ = "0.2.0"
+__author__ = "Tom Schraitle <toms@suse.de>"
